Añadir geolocalización al scrapping de depósitos

+++ src/main/java/data/model/Geolocation.java
@@ -2,15 +2,15 @@ package data.model;
 
 public class Geolocation {
 	private String type;
-	private float latitude;
-	private float longitude;
+	private double latitude;
+	private double longitude;
 	
 	public Geolocation()
 	{
 		type = "Point";
 	}
 	
-	public Geolocation(float latitude, float longitude)
+	public Geolocation(double latitude, double longitude)
 	{
 		this.type = "Point";
 		this.latitude = latitude;
@@ -21,19 +21,19 @@ public class Geolocation {
 		return type;
 	}
 
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
